load park fire centre and zone data

+++ src/backend/config/functions/bootstrap.js
@@ -186,130 +186,6 @@ const loadProtectedLandData = async (protectedLandData) => {
   } catch (error) {}
 };
 
-const loadAccessStatus = async () => {
-  const modelName = "access-status";
-  const loadSetting = await getDataLoadSetting(modelName);
-
-  if (loadSetting && loadSetting.purge)
-    await strapi.services[modelName].delete();
-
-  if (loadSetting && !loadSetting.reload) return;
-
-  const currentData = await strapi.services[modelName].find();
-  if (currentData.length == 0) {
-    strapi.log.info("Loading Access Statuses..");
-    var jsonData = fs.readFileSync("./data/access-status.json", "utf8");
-    const dataSeed = JSON.parse(jsonData);
-    dataSeed.forEach((data) => {
-      strapi.services[modelName].create(data);
-    });
-  }
-};
-
-const loadAdvisoryStatus = async () => {
-  const modelName = "advisory-status";
-  const loadSetting = await getDataLoadSetting(modelName);
-
-  if (loadSetting && loadSetting.purge)
-    await strapi.services[modelName].delete();
-
-  if (loadSetting && !loadSetting.reload) return;
-
-  const currentData = await strapi.services[modelName].find();
-  if (currentData.length == 0) {
-    strapi.log.info(`loading ${modelName}...`);
-    var jsonData = fs.readFileSync("./data/advisory-status.json", "utf8");
-    const dataSeed = JSON.parse(jsonData);
-    dataSeed.forEach((data) => {
-      strapi.services[modelName].create(data);
-    });
-  }
-};
-
-const loadEventType = async () => {
-  const modelName = "event-type";
-  const currentData = await strapi.services[modelName].find();
-  if (currentData.length == 0) {
-    strapi.log.info("Loading Event Statuses..");
-    var jsonData = fs.readFileSync("./data/event-type.json", "utf8");
-    const dataSeed = JSON.parse(jsonData);
-    dataSeed.forEach((data) => {
-      strapi.services[modelName].create(data);
-    });
-  }
-};
-
-const loadFireCentre = async () => {
-  const currentData = await strapi.services["fire-centre"].find();
-  if (currentData.length == 0) {
-    strapi.log.info("loading fire centre...");
-    var jsonData = fs.readFileSync("./data/fire-centre.json", "utf8");
-    const dataSeed = JSON.parse(jsonData)["fire-centre"];
-    dataSeed.forEach((data) => {
-      strapi.services["fire-centre"].create(data);
-    });
-  }
-};
-
-const loadFireZone = async () => {
-  const currentData = await strapi.services["fire-zone"].find();
-  if (currentData.length == 0) {
-    strapi.log.info("loading fire zone...");
-    var jsonData = fs.readFileSync("./data/fire-zone.json", "utf8");
-    const dataSeed = JSON.parse(jsonData)["fire-zone"];
-    dataSeed.forEach(async (data) => {
-      data.fire_centre = await strapi.services["fire-centre"].findOne({
-        FireCentreNumber: data.FireCentreNumber,
-      });
-      console.log(data);
-      strapi.services["fire-zone"].create(data);
-    });
-  }
-};
-
-const loadParkFireZone = async () => {
-  strapi.log.info("loading park fire zone...");
-  var jsonData = fs.readFileSync("./data/park-fire-zone-xref.json", "utf8");
-  const dataSeed = JSON.parse(jsonData)["park-fire-zone-xref"];
-  dataSeed.forEach(async (data) => {
-    const protectedArea = await strapi.services["protected-area"].findOne({
-      ORCS: data.ORCS,
-    });
-    if (protectedArea) {
-      const fireZones = await strapi.services["fire-zone"].find({
-        FireZoneNumber: data.FireZoneNumber,
-      });
-
-      protectedArea.fire_zones = fireZones;
-
-      strapi
-        .query("protected-area")
-        .update({ id: protectedArea.id }, protectedArea);
-    }
-  });
-};
-
-const loadParkFogZone = async () => {
-  strapi.log.info("loading park fog zone...");
-  var jsonData = fs.readFileSync("./data/park-fog-zone-xref.json", "utf8");
-  const dataSeed = JSON.parse(jsonData)["park-fog-zone-xref"];
-  dataSeed.forEach(async (data) => {
-    const protectedArea = await strapi.services["protected-area"].findOne({
-      ORCS: data.ORCS,
-    });
-    if (protectedArea) {
-      if (data.FogZone === "Y") protectedArea.FogZone = true;
-      else protectedArea.FogZone = false;
-
-      strapi
-        .query("protected-area")
-        .update({ id: protectedArea.id }, protectedArea);
-    }
-  });
-};
-
-// const loadParkFireZone = (async() = {});
-
 const createApiUser = async () => {
   const authRole = await findRole("authenticated");
   const password = await strapi.admin.services.auth.hashPassword(
@@ -339,46 +215,26 @@ const createApiToken = async () => {
   );
 };
 
-const loadUrgency = async () => {
-  const modelName = "urgency";
-  const loadSetting = await getDataLoadSetting(modelName);
-
-  //if (loadSetting && loadSetting.purge)
-  await strapi.services["urgency"].delete();
-
-  //if (loadSetting && !loadSetting.reload) return;
-
-  const currentData = await strapi.services["urgency"].find();
-  if (currentData.length == 0) {
-    strapi.log.info(`loading ${modelName}...`);
-    var jsonData = fs.readFileSync("./data/urgency.json", "utf8");
-    const dataSeed = JSON.parse(jsonData);
-    dataSeed.forEach((data) => {
-      strapi.services["urgency"].create(data);
-    });
-  }
-};
-
 const loadPublicAdvisory = async () => {
   const modelName = "public-advisory";
   const loadSetting = await getDataLoadSetting(modelName);
 
-  // if (loadSetting && loadSetting.purge)
-  await strapi.services[modelName].delete();
+  if (loadSetting && loadSetting.purge)
+    await strapi.services[modelName].delete();
 
-  // if (loadSetting && !loadSetting.reload) return;
+  if (loadSetting && !loadSetting.reload) return;
 
   const currentData = await strapi.services[modelName].find();
   if (currentData.length === 0) {
     strapi.log.info(`loading ${modelName}...`);
     var jsonData = fs.readFileSync("./data/public-advisory.json", "utf8");
-    const dataSeed = JSON.parse(jsonData);
+    const dataSeed = JSON.parse(jsonData)[modelName];
 
     var jsonData = fs.readFileSync("./data/public-advisory-xref.json", "utf8");
-    const dataXref = JSON.parse(jsonData)[modelName];
+    const dataXref = JSON.parse(jsonData);
 
     dataSeed.map(async (data) => {
-      const orcsXref = await dataXref.public_advisory_xref.filter(
+      const orcsXref = await dataXref["public-advisory-xref"].filter(
         (x) => x.AdvisoryNumber == data.AdvisoryNumber
       );
 
@@ -441,13 +297,56 @@ const loadPublicAdvisory = async () => {
       data.ModifiedDate = data.ModifiedDate
         ? moment(data.ModifiedDate, "YYYY-MM-DD").tz("UTC").format()
         : null;
-      // console.log({
-      //   title: data.Title,
-      //   "protected_areas-count": data.protected_areas.length,
-      // });
-      await strapi.services["public-advisory"].create(data);
+      await strapi.services[modelName].create(data);
     });
   }
+};
+
+const loadParkFireZoneXref = async () => {
+  strapi.log.info("loading park fire zone xref...");
+  var jsonData = fs.readFileSync("./data/park-fire-zone-xref.json", "utf8");
+  const dataSeed = JSON.parse(jsonData)["park-fire-zone-xref"];
+  dataSeed.forEach(async (data) => {
+    console.log(data);
+  });
+
+  dataSeed.forEach(async (data) => {
+    const protectedArea = await strapi.services["protected-area"].findOne({
+      ORCS: data.ORCS,
+    });
+    if (protectedArea) {
+      const fireZones = await strapi.services["fire-zone"].find({
+        FireZoneNumber: data.FireZoneNumber,
+      });
+      if (fireZones) {
+        console.log(protectedArea.id, fireZones);
+        protectedArea.fire_zones = fireZones;
+
+        await strapi
+          .query("protected-area")
+          .update({ id: protectedArea.id }, protectedArea);
+      }
+    }
+  });
+};
+
+const loadParkFogZoneXref = async () => {
+  strapi.log.info("loading park fog zone xref...");
+  var jsonData = fs.readFileSync("./data/park-fog-zone-xref.json", "utf8");
+  const dataSeed = JSON.parse(jsonData)["park-fog-zone-xref"];
+  dataSeed.forEach(async (data) => {
+    const protectedArea = await strapi.services["protected-area"].findOne({
+      ORCS: data.ORCS,
+    });
+    if (protectedArea) {
+      if (data.FogZone === "Y") protectedArea.FogZone = true;
+      else protectedArea.FogZone = false;
+
+      strapi
+        .query("protected-area")
+        .update({ id: protectedArea.id }, protectedArea);
+    }
+  });
 };
 
 const getDataLoadSetting = async (modelName) => {
@@ -466,7 +365,7 @@ const getDataLoadSetting = async (modelName) => {
     message.purge = loadSetting.purge;
   }
 
-  strapi.log.info("load config", message);
+  strapi.log.info("pre-load config", message);
   return loadSetting;
 };
 
@@ -598,20 +497,30 @@ const loadJsonData = async (model, jsonFile, object) => {
     });
   }
 };
+
 const loadData = async () => {
   try {
     await loadParData();
-    await loadAccessStatus();
-    await loadAdvisoryStatus();
-    await loadEventType();
-    await loadFireCentre();
-    await loadFireZone();
-    await loadParkFogZone();
-    await loadUrgency();
-    await loadPublicAdvisory();
     await createApiToken();
     await loadBusinessHours();
     await loadStatutoryHolidays();
+    await loadJsonData(
+      "access-status",
+      "./data/access-status.json",
+      "access-status"
+    );
+    await loadJsonData(
+      "advisory-status",
+      "./data/advisory-status.json",
+      "advisory-status"
+    );
+    await loadJsonData("event-type", "./data/event-type.json", "event-type");
+    await loadJsonData("fire-centre", "./data/fire-centre.json", "fire-centre");
+    await loadJsonData("fire-zone", "./data/fire-zone.json", "fire-zone");
+    await loadJsonData("urgency", "./data/urgency.json", "urgency");
+    await loadPublicAdvisory();
+    await loadParkFogZoneXref();
+    await loadParkFireZoneXref();
   } catch (error) {
     strapi.log.error(error);
   }
@@ -625,6 +534,14 @@ module.exports = async () => {
     await setDefaultPermissions();
     await loadData();
   }
+  await loadJsonData(
+    "advisory-status",
+    "./data/advisory-status.json",
+    "advisory-status"
+  );
   await loadJsonData("fire-centre", "./data/fire-centre.json", "fire-centre");
-  // await loadJsonData("fire-zone", "./data/fire-zone.json","fire-zone")
+  await loadJsonData("fire-zone", "./data/fire-zone.json", "fire-zone");
+  await loadPublicAdvisory();
+  await loadParkFogZoneXref();
+  await loadParkFireZoneXref();
 };
